requirements.txt
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv
uvicorn
pydantic
httpx
requests
langchain_google_community
langchain_tavily
langchain_groq
langchain_openai
langgraph
langchain-google-community[places]
supabase

-e .